requirements.txt
appdirs==1.4.4
appium-python-client==2.0.0a0
apscheduler==3.6.1
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==2.0.1
attrs==19.1.0
autobahn==19.6.2
automat==0.7.0
cffi==1.12.3
constantly==15.1.0
cryptography==2.7
distlib==0.3.0
django==2.0.13
django-apscheduler==0.3.0
django-cors-headers==3.0.2
djangorestframework==3.9.4
djangorestframework-camel-case==1.0.3
et-xmlfile==1.0.1
faker==5.6.0
filelock==3.0.12
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.6.1
importlib-resources==2.0.1
incremental==17.5.0
jdcal==1.4.1
jmespath==0.10.0
lxml==4.6.2
mysql-connector-python==2.1.8
mysqlclient==1.3.12
openpyxl==2.6.2
pactverify==2.0.1
paramiko==2.7.2
pdfkit==0.6.1
pycparser==2.19
pyhamcrest==1.9.0
pymongo==3.11.0
pymysql==1.0.2
pypinyin==0.40.0
pytest-rerunfailures==9.1.1
pytz==2019.1
pywin32==224
shortuuid==1.0.1
six==1.12.0
sshtunnel==0.4.0
twisted==19.2.1
txaio==18.8.1
tzlocal==2.0.0
unknown==0.0.0
virtualenv==20.0.23
xlrd==1.2.0
xmind==1.2.0
zipp==1.2.0
zope.interface==4.6.0
